switch to sqLite-based FQDB
http://localhost:8082/index.html#/test1 was showing intermittent fails
on /data/getSeq with python complaints like "ValueError: Problem with
quality section". Forcing CherryPy to one thread eliminated the problem,
and made the app unusably slow. Using SeqIO.index_db exposed the sqLite
issue with being used in a different thread than the one in which the
connection was created. Solved by using the same pattern as in GeneDB,
where the sqLite connection is created anew on each use.